+++ private/PROPOSAL/crossection/parametrization/Annihilation.cxx
@@ -5,6 +5,7 @@
 
 #include "PROPOSAL/medium/Components.h"
 #include "PROPOSAL/medium/Medium.h"
+#include "PROPOSAL/math/MathMethods.h"
 #include "PROPOSAL/Constants.h"
 #include "PROPOSAL/Logging.h"
