+++ private/PROPOSAL/crossection/parametrization/EpairProduction.cxx
@@ -3,6 +3,7 @@
 
 #include "PROPOSAL/crossection/parametrization/EpairProduction.h"
 
+#include "PROPOSAL/math/MathMethods.h"
 #include "PROPOSAL/medium/Components.h"
 #include "PROPOSAL/medium/Medium.h"
