+++ private/PROPOSAL/crossection/parametrization/PhotoPairProduction.cxx
@@ -6,6 +6,7 @@
 #include "PROPOSAL/medium/Components.h"
 #include "PROPOSAL/medium/Medium.h"
 
+#include "PROPOSAL/math/MathMethods.h"
 #include "PROPOSAL/Constants.h"
 #include "PROPOSAL/Logging.h"
 #include "PROPOSAL/math/RandomGenerator.h"
