+++ private/PROPOSAL/math/MathMethods.cxx
@@ -1,12 +1,147 @@
-#include "PROPOSAL/math/MathMethods.h"
-#include <math.h>
+
+#include <cmath>
 #include <fstream>
 #include <iostream>
+#include "PROPOSAL/math/MathMethods.h"
 #include "PROPOSAL/Logging.h"
 #include "PROPOSAL/Output.h"
 #include "PROPOSAL/json.hpp"
 
 namespace PROPOSAL {
+
+double inverseErrorFunction(double p) {
+    if (p <= 0 || p >= 1) {
+        log_fatal(
+            "The inverse Error function can just handle values between 0 and "
+            "1.");
+    }
+    double a_arr[] = {-3.969683028665376e+01, 2.209460984245205e+02,
+                      -2.759285104469687e+02, 1.383577518672690e+02,
+                      -3.066479806614716e+01, 2.506628277459239e+00};
+    double b_arr[] = {-5.447609879822406e+01, 1.615858368580409e+02,
+                      -1.556989798598866e+02, 6.680131188771972e+01,
+                      -1.328068155288572e+01};
+    double c_arr[] = {-7.784894002430293e-03, -3.223964580411365e-01,
+                      -2.400758277161838e+00, -2.549732539343734e+00,
+                      4.374664141464968e+00,  2.938163982698783e+00};
+    double d_arr[] = {7.784695709041462e-03, 3.224671290700398e-01,
+                      2.445134137142996e+00, 3.754408661907416e+00};
+    double p_low = 0.02425;
+    double p_high = 1 - p_low;
+
+    double q, r, x, e, u;
+    if (p < p_low) {
+        // Rational approximation for lower region.
+        q = std::sqrt(-2 * std::log(p));
+        x = (((((c_arr[0] * q + c_arr[1]) * q + c_arr[2]) * q + c_arr[3]) * q +
+              c_arr[4]) *
+                 q +
+             c_arr[5]) /
+            ((((d_arr[0] * q + d_arr[1]) * q + d_arr[2]) * q + d_arr[3]) * q +
+             1);
+    } else if (p_low <= p && p <= p_high) {
+        // Rational approximation for central region.
+        q = p - 0.5;
+        r = q * q;
+        x = (((((a_arr[0] * r + a_arr[1]) * r + a_arr[2]) * r + a_arr[3]) * r +
+              a_arr[4]) *
+                 r +
+             a_arr[5]) *
+            q /
+            (((((b_arr[0] * r + b_arr[1]) * r + b_arr[2]) * r + b_arr[3]) * r +
+              b_arr[4]) *
+                 r +
+             1);
+    } else {
+        // Rational approximation for upper region.
+        q = std::sqrt(-2 * std::log(1 - p));
+        x = -(((((c_arr[0] * q + c_arr[1]) * q + c_arr[2]) * q + c_arr[3]) * q +
+               c_arr[4]) *
+                  q +
+              c_arr[5]) /
+            ((((d_arr[0] * q + d_arr[1]) * q + d_arr[2]) * q + d_arr[3]) * q +
+             1);
+    }
+
+    // Refining the result:
+    // One iteration of Halley’s rational method (third order)
+    // gives full machine precision.
+    e = 0.5 * std::erfc(-x / SQRT2) - p;
+    u = e * std::sqrt(2 * PI) * std::exp(0.5 * x * x);
+    x = x - u / (1 + x * u * 0.5);
+    return x;
+}
+
+// ------------------------------------------------------------------------- //
+double dilog(double x) {
+    double C_arr[] = {
+        0.42996693560813697,  0.40975987533077105,  -0.01858843665014592,
+        0.00145751084062268,  -0.00014304184442340, 0.00001588415541880,
+        -0.00000190784959387, 0.00000024195180854,  -0.00000003193341274,
+        0.00000000434545063,  -0.00000000060578480, 0.00000000008612098,
+        -0.00000000001244332, 0.00000000000182256,  -0.00000000000027007,
+        0.00000000000004042,  -0.00000000000000610, 0.00000000000000093,
+        -0.00000000000000014, 0.00000000000000002};
+
+    double HF = 0.5;
+
+    if (x == 1)
+        return PI * PI / 6.0;
+    else if (x == -1)
+        return -PI * PI / 12.0;
+    else {
+        double _T = -x;
+        double Y, S, A, B0, B1, B2, _H, _ALFA;
+
+        if (_T <= -2.0) {
+            Y = -1.0 / (1.0 + _T);
+            S = 1.0;
+            B1 = std::log(-_T);
+            B2 = std::log(1.0 + 1.0 / _T);
+            A = -PI * PI / 3.0 + HF * (B1 * B1 - B2 * B2);
+        } else if (_T < -1.0) {
+            Y = -1.0 - _T;
+            S = -1.0;
+            A = std::log(-_T);
+            A = -PI * PI / 6.0 + A * (A + std::log(1 + 1 / _T));
+        } else if (_T <= -0.5) {
+            Y = -(1 + _T) / _T;
+            S = 1.0;
+            A = std::log(-_T);
+            A = -PI * PI / 6.0 + A * (-HF * A + std::log(1.0 + _T));
+        } else if (_T < 0.0) {
+            Y = -_T / (1 + _T);
+            S = -1.0;
+            A = std::log(1 + _T);
+            A = HF * A * A;
+        } else if (_T <= 1.0) {
+            Y = _T;
+            S = 1.0;
+            A = 0.0;
+        } else {
+            Y = 1.0 / _T;
+            S = -1.0;
+            A = std::log(_T);
+            A = PI * PI / 6 + HF * A * A;
+        }
+
+        _H = Y + Y - 1;
+        _ALFA = _H + _H;
+        B1 = 0.0;
+        B2 = 0.0;
+
+        for (int i = 19; i >= 0.0; --i) {
+            B0 = C_arr[i] + _ALFA * B1 - B2;
+            B2 = B1;
+            B1 = B0;
+        }
+
+        return -(S * (B0 - _H * B2) + A);
+    }
+
+    return x;
+}
+
 
 double NewtonRaphson(std::function<double(double)> func,
                      std::function<double(double)> dfunc,
@@ -51,7 +186,7 @@ double NewtonRaphson(std::function<double(double)> func,
     }
 
     rts = xinit;  // initial guess for root
-    dxold = fabs(x2 - x1);
+    dxold = std::abs(x2 - x1);
     dx = dxold;
 
     f = func(rts);
@@ -59,7 +194,7 @@ double NewtonRaphson(std::function<double(double)> func,
 
     for (int j = 1; j < MAX_STEPS; j++) {
         if ((((rts - xh) * df - f) * ((rts - xl) * df - f) > 0.0) ||
-            (fabs(2.0 * f) > fabs(dxold * df))) {
+            (std::abs(2.0 * f) > std::abs(dxold * df))) {
             // use bisection if Newton method produces an x that is out of range
             // or if Newton method is not decreasing fast enough
             dxold = dx;
@@ -79,7 +214,7 @@ double NewtonRaphson(std::function<double(double)> func,
             }
         }
 
-        if (fabs(dx) < xacc) {
+        if (std::abs(dx) < xacc) {
             // convergence criterion reached
             return rts;
         }
