+++ private/python/medium.cxx
@@ -1,4 +1,9 @@
-#include "PROPOSAL/PROPOSAL.h"
+
+#include "PROPOSAL/medium/Components.h"
+#include "PROPOSAL/medium/Medium.h"
+#include "PROPOSAL/medium/density_distr/density_exponential.h"
+#include "PROPOSAL/medium/density_distr/density_polynomial.h"
+#include "PROPOSAL/medium/density_distr/density_splines.h"
 #include "pyBindings.h"
 
 #define MEDIUM_DEF(module, cls)                                 \
