+++ public/PROPOSAL/Sector.h
